move BoP to own statistic
